+++ flink-libraries/flink-gelly/src/test/java/org/apache/flink/graph/generator/CycleGraphTest.java
@@ -46,7 +46,7 @@ public class CycleGraphTest extends GraphGeneratorTestBase {
 
 	@Test
 	public void testGraphMetrics() throws Exception {
-		int vertexCount = 100;
+		long vertexCount = 100;
 
 		Graph<LongValue, NullValue, NullValue> graph = new CycleGraph(env, vertexCount)
 			.generate();
